Update documentation pages sidebar
